+++ src/collections/domain/repositories/ICollectionsRepository.ts
@@ -14,6 +14,7 @@ import { CollectionSummary } from '../models/CollectionSummary'
 import { AllowedStorageDrivers } from '../models/AllowedStorageDrivers'
 import { StorageDriver } from '../../../core/domain/models/StorageDriver'
 import { LinkingObjectType } from '../useCases/GetCollectionsForLinking'
+import { RoleAlias } from '../../../roles/domain/models/RoleAlias'
 
 export interface ICollectionsRepository {
   getCollection(collectionIdOrAlias: number | string): Promise<Collection>
@@ -39,6 +40,10 @@ export interface ICollectionsRepository {
   getCollectionUserPermissions(
     collectionIdOrAlias: number | string
   ): Promise<CollectionUserPermissions>
+  setDefaultContributorRole(
+    collectionIdOrAlias: number | string,
+    roleAlias: RoleAlias | string
+  ): Promise<void>
   getCollectionItems(
     collectionId?: string,
     limit?: number,